Update CI image and supported compilers (#2611)
Update CI image to Ubuntu 23.10 and with that update supported compilers to GCC 11+ and LLVM 15+.  Includes further changes to Jenkins, much clean up for newer clang-tidy, and more.

+++ src/lib/scheduler/task_queue.hpp
@@ -7,7 +7,7 @@
 #include <condition_variable>
 #include <memory>
 
-#include <tbb/concurrent_queue.h>  // NOLINT(build/include_order): wronlgy identified as a C header.
+#include <oneapi/tbb/concurrent_queue.h>  // NOLINT(build/include_order): cpplint identifies TBB as C system headers.
 
 #include "concurrentqueue.h"  // The lightweight semaphore uses definitions of concurrentqueue.h.
 #include "lightweightsemaphore.h"
@@ -57,7 +57,7 @@ class TaskQueue {
    */
   size_t estimate_load() const;
 
-  void signal(const size_t count);
+  void signal(const int32_t count);
 
   /**
    * Semaphore to signal waiting workers for new tasks.
